Upgrade armon/go-metrics to hashicorp/go-metrics
Also fixup the library to remove the dual v1 & v2 paths. Now the main branch will support only v3.

v3 is API compatible and on-disk-storage compatible with v2. The main difference is that it requires the application to be using hashicorp/go-metrics instead of armon/go-metrics in order for metrics to be emitted properly.
