+++ bolt_store.go
@@ -4,9 +4,9 @@ import (
 	"errors"
 	"time"
 
-	metrics "github.com/armon/go-metrics"
-	"github.com/boltdb/bolt"
+	metrics "github.com/hashicorp/go-metrics"
 	"github.com/hashicorp/raft"
+	"go.etcd.io/bbolt"
 )
 
 const (
@@ -24,25 +24,25 @@ var (
 	ErrKeyNotFound = errors.New("not found")
 )
 
-// BoltStore provides access to BoltDB for Raft to store and retrieve
+// BoltStore provides access to Bbolt for Raft to store and retrieve
 // log entries. It also provides key/value storage, and can be used as
 // a LogStore and StableStore.
 type BoltStore struct {
 	// conn is the underlying handle to the db.
-	conn *bolt.DB
+	conn *bbolt.DB
 
 	// The path to the Bolt database file
 	path string
 }
 
-// Options contains all the configuration used to open the BoltDB
+// Options contains all the configuration used to open the Bbolt
 type Options struct {
-	// Path is the file path to the BoltDB to use
+	// Path is the file path to the Bbolt to use
 	Path string
 
-	// BoltOptions contains any specific BoltDB options you might
+	// BoltOptions contains any specific Bbolt options you might
 	// want to specify [e.g. open timeout]
-	BoltOptions *bolt.Options
+	BoltOptions *bbolt.Options
 
 	// NoSync causes the database to skip fsync calls after each
 	// write to the log. This is unsafe, so it should be used
@@ -62,10 +62,10 @@ func NewBoltStore(path string) (*BoltStore, error) {
 	return New(Options{Path: path})
 }
 
-// New uses the supplied options to open the BoltDB and prepare it for use as a raft backend.
+// New uses the supplied options to open the Bbolt and prepare it for use as a raft backend.
 func New(options Options) (*BoltStore, error) {
 	// Try to connect
-	handle, err := bolt.Open(options.Path, dbFileMode, options.BoltOptions)
+	handle, err := bbolt.Open(options.Path, dbFileMode, options.BoltOptions)
 	if err != nil {
 		return nil, err
 	}
@@ -107,6 +107,10 @@ func (b *BoltStore) initialize() error {
 	return tx.Commit()
 }
 
+func (b *BoltStore) Stats() bbolt.Stats {
+	return b.conn.Stats()
+}
+
 // Close is used to gracefully close the DB connection.
 func (b *BoltStore) Close() error {
 	return b.conn.Close()
@@ -144,8 +148,10 @@ func (b *BoltStore) LastIndex() (uint64, error) {
 	}
 }
 
-// GetLog is used to retrieve a log from BoltDB at a given index.
+// GetLog is used to retrieve a log from Bbolt at a given index.
 func (b *BoltStore) GetLog(idx uint64, log *raft.Log) error {
+	defer metrics.MeasureSince([]string{"raft", "boltdb", "getLog"}, time.Now())
+
 	tx, err := b.conn.Begin(false)
 	if err != nil {
 		return err
@@ -169,6 +175,7 @@ func (b *BoltStore) StoreLog(log *raft.Log) error {
 // StoreLogs is used to store a set of raft logs
 func (b *BoltStore) StoreLogs(logs []*raft.Log) error {
 	now := time.Now()
+
 	tx, err := b.conn.Begin(true)
 	if err != nil {
 		return err
