+++ go.mod
@@ -1,10 +1,11 @@
-module github.com/hashicorp/raft-boltdb
+module github.com/hashicorp/raft-boltdb/v3
 
 go 1.16
 
 require (
-	github.com/armon/go-metrics v0.3.8 // indirect
 	github.com/boltdb/bolt v1.3.1
+	github.com/hashicorp/go-metrics v0.5.1
 	github.com/hashicorp/go-msgpack v0.5.5
 	github.com/hashicorp/raft v1.1.0
+	go.etcd.io/bbolt v1.3.5
 )
