--- v2/go.mod
@@ -1,12 +1,0 @@
-module github.com/hashicorp/raft-boltdb/v2
-
-go 1.16
-
-require (
-	github.com/armon/go-metrics v0.0.0-20190430140413-ec5e00d3c878
-	github.com/boltdb/bolt v1.3.1
-	github.com/hashicorp/go-msgpack v0.5.5
-	github.com/hashicorp/raft v1.1.0
-	github.com/hashicorp/raft-boltdb v0.0.0-20210409134258-03c10cc3d4ea
-	go.etcd.io/bbolt v1.3.5
-)
